a function that returns the number of elements in a linked list

+++ 0x13-more_singly_linked_lists/1-listint_len.c
@@ -9,13 +9,13 @@
 
 size_t listint_len(const listint_t *h)
 {
-	int s = 0;
+size_t s = 0;
 
-	while (h != NULL)
-	{
-		h = h->next;
-		s++;
-	}
+while (h != NULL)
+{
+	h = h->next;
+	s++;
+}
 
-	return (s);
+return (s);
 }
